+++ libs/angular-vanilla/src/lib/components/currency/currency.component.ts
@@ -1,0 +1,36 @@
+import { CommonModule } from '@angular/common';
+import { Component, CUSTOM_ELEMENTS_SCHEMA, inject, OnDestroy, OnInit } from '@angular/core';
+import * as Angular from '@golemui/angular';
+import * as Core from '@golemui/core';
+import { CurrencyProps } from '@golemui/shared-vanilla';
+
+@Component({
+  standalone: true,
+  selector: 'gui-currency-control',
+  imports: [CommonModule],
+  providers: [Angular.ControlFieldAdapter],
+  templateUrl: './currency.component.html',
+  host: {
+    class: 'gui-currency',
+  },
+  schemas: [CUSTOM_ELEMENTS_SCHEMA],
+})
+export class CurrencyComponent implements OnInit, OnDestroy, Core.WithField {
+  field!: Core.ControlField<number>;
+  protected adapter: Angular.ControlFieldAdapter<number, CurrencyProps> = inject(
+    Angular.ControlFieldAdapter,
+  );
+
+  ngOnInit(): void {
+    this.adapter.init(this.field);
+  }
+
+  ngOnDestroy(): void {
+    this.adapter.destroy();
+  }
+
+  valueChanged(event: Event) {
+    const value = (event as CustomEvent).detail.value;
+    this.adapter.valueChanged(value);
+  }
+}