+++ libs/react-vanilla/CHANGELOG.md
@@ -1,3 +1,18 @@
+## 0.18.0 (2026-01-13)
+
+### 🚀 Features
+
+- add currency field ([232ab3a](https://github.com/golemui/formforge/commit/232ab3a))
+
+### 🩹 Fixes
+
+- **react:** fix Field Function support ([09f86fa](https://github.com/golemui/formforge/commit/09f86fa))
+
+### ❤️ Thank You
+
+- Mud Scientist
+- Raul Jimenez @Elecash
+
 ## 0.17.0 (2026-01-11)
 
 ### 🚀 Features